- because the feed I am using, doesn't return the state when it returns a list of matching cities, then when you enter `Portland` as your city and Portland (Maine) and Portland (Oregon) are returned, there is no visual distinction. Now I am displaying the lat/long in the city PopupMenu as well, to at least be able to distinguish.
- also fixing a bug where we weren't `clearing` the `PopupMenu` each time, so new sets of matching cities were being added onto old ones vs. replacing them.

+++ app/src/main/java/com/premise/weather/models/Location.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Location(
     @SerializedName("woeid")
-    val locationId: Long,
+    val locationId: Int,
     val title: String,
     val location_type: String,
     val latt_long: String,
