+++ app/src/main/java/com/premise/weather/networking/WeatherRepository.kt
@@ -4,11 +4,11 @@ import com.premise.weather.models.Weather
 import com.premise.weather.networking.api.WeatherApi
 import retrofit2.Call
 
-object WeatherRepository : BaseRepository<WeatherApi, Long, Weather>(WeatherApi::class.java) {
+object WeatherRepository : BaseRepository<WeatherApi, Int, Weather>(WeatherApi::class.java) {
 
-    fun getWeather(locationId: Long) = fetch(locationId)
+    fun getWeather(locationId: Int) = fetch(locationId)
 
-    override fun apiCall(api: WeatherApi, param: Long): Call<Weather> {
+    override fun apiCall(api: WeatherApi, param: Int): Call<Weather> {
         return api.getLocations(param)
     }
 }
